move (de)serialization rules to separate file
