+++ crates/crypto/src/header/file.rs
@@ -2,9 +2,12 @@ use std::io::{Read, Seek};
 
 use crate::{
 	error::Error,
-	keys::hashing::Params,
 	primitives::{Algorithm, HashingAlgorithm, Mode, ENCRYPTED_MASTER_KEY_LEN, SALT_LEN},
 };
+
+
+// random values, can be changed
+pub const MAGIC_BYTES: [u8; 6] = [0x08, 0xFF, 0x55, 0x32, 0x58, 0x1A];
 
 // Everything contained within this header can be flaunted around with minimal security risk
 // The only way this could compromise any data is if a weak password/key was used
@@ -40,65 +43,10 @@ pub enum FileHeaderVersion {
 	V1,
 }
 
-// TODO(brxken128): maybe use a deserialization error
-// TODO(brxken128): move all serialization/deserialization rules
-impl FileHeaderVersion {
-	#[must_use]
-	pub const fn serialize(&self) -> [u8; 2] {
-		match self {
-			Self::V1 => [0x0A, 0x01],
-		}
-	}
-
-	pub const fn deserialize(bytes: [u8; 2]) -> Result<Self, Error> {
-		match bytes {
-			[0x0A, 0x01] => Ok(Self::V1),
-			_ => Err(Error::FileHeader),
-		}
-	}
-}
-
 pub enum FileKeyslotVersion {
 	V1,
 }
 
-impl FileKeyslotVersion {
-	#[must_use]
-	pub const fn serialize(&self) -> [u8; 2] {
-		match self {
-			Self::V1 => [0x0D, 0x01],
-		}
-	}
-
-	pub const fn deserialize(bytes: [u8; 2]) -> Result<Self, Error> {
-		match bytes {
-			[0x0D, 0x01] => Ok(Self::V1),
-			_ => Err(Error::FileHeader),
-		}
-	}
-}
-
-impl HashingAlgorithm {
-	#[must_use]
-	pub const fn serialize(&self) -> [u8; 2] {
-		match self {
-			Self::Argon2id(p) => match p {
-				Params::Standard => [0x0F, 0x01],
-				Params::Hardened => [0x0F, 0x02],
-				Params::Paranoid => [0x0F, 0x03],
-			},
-		}
-	}
-
-	pub const fn deserialize(bytes: [u8; 2]) -> Result<Self, Error> {
-		match bytes {
-			[0x0F, 0x01] => Ok(Self::Argon2id(Params::Standard)),
-			[0x0F, 0x02] => Ok(Self::Argon2id(Params::Hardened)),
-			[0x0F, 0x03] => Ok(Self::Argon2id(Params::Paranoid)),
-			_ => Err(Error::FileHeader),
-		}
-	}
-}
 
 impl FileKeyslot {
 	#[must_use]
@@ -169,45 +117,6 @@ impl FileKeyslot {
 		}
 	}
 }
-
-impl Algorithm {
-	#[must_use]
-	pub const fn serialize(&self) -> [u8; 2] {
-		match self {
-			Self::XChaCha20Poly1305 => [0x0B, 0x01],
-			Self::Aes256Gcm => [0x0B, 0x02],
-		}
-	}
-
-	pub const fn deserialize(bytes: [u8; 2]) -> Result<Self, Error> {
-		match bytes {
-			[0x0B, 0x01] => Ok(Self::XChaCha20Poly1305),
-			[0x0B, 0x02] => Ok(Self::Aes256Gcm),
-			_ => Err(Error::FileHeader),
-		}
-	}
-}
-
-impl Mode {
-	#[must_use]
-	pub const fn serialize(&self) -> [u8; 2] {
-		match self {
-			Self::Stream => [0x0C, 0x01],
-			Self::Memory => [0x0C, 0x02],
-		}
-	}
-
-	pub const fn deserialize(bytes: [u8; 2]) -> Result<Self, Error> {
-		match bytes {
-			[0x0C, 0x01] => Ok(Self::Stream),
-			[0x0C, 0x02] => Ok(Self::Memory),
-			_ => Err(Error::FileHeader),
-		}
-	}
-}
-
-// random values, can be changed
-pub const MAGIC_BYTES: [u8; 6] = [0x08, 0xFF, 0x55, 0x32, 0x58, 0x1A];
 
 impl FileHeader {
 	#[must_use]
